pyproject.toml
[project]
name = "spl-access-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.1",
    "schedule>=1.2.2",
]
